+++ spec/models/person_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Person, type: :model, versioning: true do
   end
 
   describe "#cars and bicycles" do
-    it "can be reified", skip: "Known Issue #594" do
+    it "can be reified" do
       person = Person.create(name: "Frank")
       car = Car.create(name: "BMW 325")
       bicycle = Bicycle.create(name: "BMX 1.0")
@@ -23,11 +23,16 @@ RSpec.describe Person, type: :model, versioning: true do
 
       expect(person.reload.versions.length).to(eq(3))
 
-      # Will fail because the correct sub type of the STI model is not present at query.
       # See https://github.com/airblade/paper_trail/issues/594
-      second_version = person.reload.versions.second.reify(has_one: true)
-      expect(second_version.car.name).to(eq("BMW 325"))
-      expect(second_version.bicycle.name).to(eq("BMX 1.0"))
+      expect {
+        person.reload.versions.second.reify(has_one: true)
+      }.to(
+        raise_error(::PaperTrail::Reifiers::HasOne::FoundMoreThanOne) do |err|
+          expect(err.message.squish).to match(
+            /Expected to find one Vehicle, but found 2/
+          )
+        end
+      )
     end
   end
 end